refactor(header): decompose directory - create components directory - move HeaderButtons, HeaderSearch into components - create barrel file - fix buttonLink name selector
