+++ src/components/Header/Header.module.css
@@ -26,73 +26,6 @@
   cursor: pointer;
 }
 
-.searchHolder {
-  width: 600px;
-  position: relative;
-}
-
-.search {
-  width: 100%;
-  padding: 0.65rem 2.5rem;
-  border: 1px solid #bbbbbb;
-  border-radius: 8px;
-  font-size: 16px;
-  color: var(--color-text-secondary);
-  text-overflow: ellipsis;
-}
-
-.search:focus {
-  outline: 1px solid var(--color-primary);
-  border: 1px solid var(--color-primary);
-}
-
-.searchIcon,
-.clearButton {
-  position: absolute;
-  top: 10px;
-}
-
-.searchIcon {
-  left: 10px;
-}
-
-.clearButton {
-  right: 10px;
-}
-
-.buttons {
-  display: flex;
-  align-items: center;
-  gap: 1.25rem;
-}
-
-.buttonLink {
-  padding: 0.5rem;
-  border-radius: 50%;
-  transition: background-color 250ms ease-in-out;
-  position: relative;
-}
-
-.button:hover {
-  background-color: #eeeeee;
-}
-
-.badge {
-  position: absolute;
-  top: 3px;
-  right: 0;
-  background-color: #ef4444;
-  color: white;
-  border-radius: 50%;
-  min-width: 20px;
-  height: 20px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  font-size: 12px;
-  font-weight: bold;
-}
-
 .menuToggle {
   display: none;
   flex-direction: column;
