+++ src/components/Header/components/HeaderButtons/HeaderButtons.module.css
@@ -1,0 +1,38 @@
+.buttons {
+  display: flex;
+  align-items: center;
+  gap: 1.25rem;
+}
+
+.buttonLink {
+  padding: 0.5rem;
+  border-radius: 50%;
+  transition: background-color 250ms ease-in-out;
+  position: relative;
+}
+
+.buttonLink:hover {
+  background-color: #eeeeee;
+}
+
+.badge {
+  position: absolute;
+  top: 3px;
+  right: 0;
+  background-color: #ef4444;
+  color: white;
+  border-radius: 50%;
+  min-width: 20px;
+  height: 20px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  font-size: 12px;
+  font-weight: bold;
+}
+
+@media (max-width: 600px) {
+  .buttons {
+    display: none;
+  }
+}
