+++ src/components/Header/components/HeaderSearch/HeaderSearch.module.css
@@ -1,0 +1,33 @@
+.searchHolder {
+  width: 600px;
+  position: relative;
+}
+
+.search {
+  width: 100%;
+  padding: 0.65rem 2.5rem;
+  border: 1px solid #bbbbbb;
+  border-radius: 8px;
+  font-size: 16px;
+  color: var(--color-text-secondary);
+  text-overflow: ellipsis;
+}
+
+.search:focus {
+  outline: 1px solid var(--color-primary);
+  border: 1px solid var(--color-primary);
+}
+
+.searchIcon,
+.clearButton {
+  position: absolute;
+  top: 10px;
+}
+
+.searchIcon {
+  left: 10px;
+}
+
+.clearButton {
+  right: 10px;
+}
