+++ src/components/Header/components/index.ts
@@ -1,0 +1,2 @@
+export { default as HeaderButtons } from './HeaderButtons/HeaderButtons';
+export { default as HeaderSearch } from './HeaderSearch/HeaderSearch';
